Refactor Node.expand to OrientedBoundingBox.expand
